Standard tiling + fixed compilation errors. Expression need to be develpt and execute function doesn't work with negative values.

+++ Multiplication.h
@@ -8,7 +8,7 @@
 class Multiplication
 {
     public:
-        virtual vector <MultiplicationTree> dispositions (short, short) = 0;
+        virtual vector <MultiplicationTree> dispositions (int, int) = 0;
         virtual ~Multiplication () = default;
 };
